增加否定以及双重否定判断 优化get_lower 增加配置文件内容：是否输出code为0的事件 针对颁布文件和target意义重复的问题对上一次的修改作出调整（并发现版本有问题？）

+++ main.py
@@ -11,14 +11,28 @@ import time
 from FromCorenlpConverter import FromCorenlpConverter
 from petrarch2.petrarch2 import main as petrarch2_main
 
+# iface参数指Linux的网卡接口，如(eth0,wlan0)，这个参数只支持Linux并且需要root权限
+def get_free_port(iface=None):
+    s = socket.socket()
+
+    if iface:
+        s.setsockopt(socket.SOL_SOCKET, 25, bytes(iface, 'utf8'))
+
+    s.bind(('', 0))
+    port = s.getsockname()[1]
+    s.close()
+
+    return port
 
 def wait_process(port):
-    ret = os.popen("netstat -nao|findstr " + str(port))
-    str_list = ret.read().decode('gbk')
-
-    ret_list = re.split('', str_list)
+    cmd_check = "lsof -i:" + str(port) if env == 0 else "netstat -nao|findstr " + str(port)
+    ret = os.popen(cmd_check)
     try:
-        process_pid = list(ret_list[0].split())[-1]
+        # 注意解码方式和cmd要相同，即为"gbk"，否则输出乱码
+        str_list = ret.readlines()[-1].decode('gbk')
+
+        ret_list = re.split('', str_list)
+        process_pid = list(ret_list[0].split())[1]
         if process_pid:
             return True
         else:
@@ -30,14 +44,17 @@ def wait_process(port):
         return False
 
 def kill_process(port):
-    ret = os.popen("netstat -nao|findstr " + str(port))
-    # 注意解码方式和cmd要相同，即为"gbk"，否则输出乱码
-    str_list = ret.read().decode('gbk')
-
-    ret_list = re.split('', str_list)
+    cmd_check = "lsof -i:" + str(port) if env == 0 else "netstat -nao|findstr " + str(port)
+    ret = os.popen(cmd_check)
     try:
-        process_pid = list(ret_list[0].split())[-1]
-        os.popen('taskkill /pid ' + str(process_pid) + ' /F')
+        # 注意解码方式和cmd要相同，即为"gbk"，否则输出乱码
+        str_list = ret.readlines()[-1].decode('gbk')
+
+        ret_list = re.split('', str_list)
+        process_pid = list(ret_list[0].split())[1]
+
+        cmd_kill='kill -9 ' + str(process_pid) if env == 0 else 'taskkill /pid ' + str(process_pid) + ' /F'
+        os.popen(cmd_kill)
         print ("端口已被释放")
     except:
         print ("端口未被使用")
@@ -48,41 +65,36 @@ if __name__ == "__main__":
     # If you are using python2, the first two lines are needed.
     reload(sys)
     sys.setdefaultencoding('utf-8')
-    xml_file_name = 'format_text'
-    input_path = 'input/'
-    input_name = 'raw.txt'
-    xml_output_path = 'output/'
-    corenlp_path = 'stanford-corenlp-full-2018-10-05'
-    port = 6330
-
-    kill_process(port)
-    while(wait_process(port)):
-        print('端口一直被占用中,30秒后自动检测')
-        time.sleep(30)
-
-
-    output_path = './'
-    output_filename = 'events.txt'
-
-    format_text = 'format_text'
-
-
-
-    # reader.parse_Config()
-    # if not gcp.input_path == "":
-    #     input_path = gcp.input_path
-    # if not gcp.input_name == "":
-    #     input_name = gcp.input_name
-    # if not gcp.xml_output_path == "":
-    #     xml_output_path = gcp.xml_output_path
-    # if not gcp.output_path == "":
-    #     output_path = gcp.output_path
-    # if not gcp.output_name == "":
-    #     output_filename = gcp.output_name
-
-
-
-
+    env = 1
+    print(os.getcwd())
+    reader.parse_Config()
+    input_path = input_name = xml_output_path = output_path = output_filename = xml_file_name = format_text = corenlp_path = ""
+    port = -1
+    if not gcp.input_path == "":
+        input_path = gcp.input_path
+    if not gcp.input_name == "":
+        input_name = gcp.input_name
+    if not gcp.xml_output_path == "":
+        xml_output_path = gcp.xml_output_path
+    if not gcp.output_path == "":
+        output_path = gcp.output_path
+    if not gcp.output_name == "":
+        output_filename = gcp.output_name
+    if not gcp.xml_file_name == "":
+        xml_file_name = gcp.xml_file_name
+    if not gcp.format_text == "":
+        format_text = gcp.format_text
+    if not gcp.corenlp_path == "":
+        corenlp_path = gcp.corenlp_path
+    if not gcp.port == -1:#自选端口
+        port = gcp.port
+        kill_process(port)
+        while (wait_process(port)):
+            print('端口一直被占用中,30秒后自动检测')
+            time.sleep(30)
+    else:
+        port = get_free_port()
+        print("程序将在"+str(port)+"号端口启动")
 
     formatted_lines = []
     with open(input_path+input_name , 'r') as fp:
@@ -94,11 +106,11 @@ if __name__ == "__main__":
     with open(input_path + format_text + '.txt', 'w') as fw:
         fw.writelines(formatted_lines)
 
-    if os.path.exists('evts.test.txt'):
-        with open('evts.test.txt', 'w') as fw:
+    if os.path.exists(output_filename):
+        with open(output_filename, 'w') as fw:
             fw.write('')
 
-    flag = True
+    flag = False
     if flag:
         converter = FromCorenlpConverter(input_path + format_text + '.txt', '', corenlp_path, port)
